+++ lib/cached/md5_cached.cmake
@@ -1,5 +1,5 @@
 set(lexer_lxx_md5_cached "29c69cfc368018c0bf8fe3d63f00c677")
-set(parser_yxx_md5_cached "19feeb447b96bc8b39ce497ab45af0ae")
+set(parser_yxx_md5_cached "2b54a2ab1b23fe6a2179d0d3e6d23d03")
 set(regex_lexer_lxx_md5_cached "d4bdc40f824511fb088b32f57405136c")
 set(regex_parser_yxx_md5_cached "e44695248dcfad64a56982f126f8e4f5")
 set(ts_minizinc_grammar_js_md5_cached "4b3dbfa9ad19fcdd41012baee4064d04")